+++ com.vorwardit.wordsworthcms/API/PageLayoutController.cs
@@ -3,6 +3,7 @@ using com.vorwardit.wordsworthcms.BusinessLogic.Interfaces;
 using com.vorwardit.wordsworthcms.Engine;
 using com.vorwardit.wordsworthcms.Models;
 using com.vorwardit.wordsworthcms.ViewModels;
+using Microsoft.AspNet.Identity;
 using System;
 using System.Collections.Generic;
 using System.Data.Entity;
@@ -19,16 +20,18 @@ namespace com.vorwardit.wordsworthcms.API
     public class PageLayoutController : ApiController
     {
         IPageLayoutService pageLayoutService;
+		IUserService userService;
 
-        public PageLayoutController(IPageLayoutService pageLayoutService)
+		public PageLayoutController(IPageLayoutService pageLayoutService, IUserService userService)
         {
             this.pageLayoutService = pageLayoutService;
+			this.userService = userService;
         }
 
         [HttpGet]
         public async Task<IHttpActionResult> GetBySiteId(Guid siteId)
         {
-            var pageLayouts = await pageLayoutService.GetBySiteIdAsync(siteId);
+			var pageLayouts = await pageLayoutService.GetBySiteIdAsync(siteId);
             return Ok(from l in pageLayouts
                       select new
                       {
@@ -41,13 +44,19 @@ namespace com.vorwardit.wordsworthcms.API
         [HttpGet]
         public async Task<IHttpActionResult> Get(Guid id)
         {
-            return Ok(Mapper.Map<PageLayoutViewModel>(await pageLayoutService.GetAsync(id)));
+			return Ok(Mapper.Map<PageLayoutViewModel>(await pageLayoutService.GetAsync(id)));
         }
 
         [HttpPost]
         public async Task<IHttpActionResult> Post(PageLayoutViewModel model)
         {
-            PageLayout pageLayout = await pageLayoutService.GetOrCreateAsync(model.PageLayoutId);
+			var allowed = await userService.HasPermission(User.Identity.GetUserId(), PermissionType.Designer, null);
+			if (!allowed)
+			{
+				return BadRequest("User is not authorised to perform this action");
+			}
+
+			PageLayout pageLayout = await pageLayoutService.GetOrCreateAsync(model.PageLayoutId);
             pageLayout.LayoutId = model.LayoutId;
             pageLayout.Name = model.Name;
             pageLayout.Body = model.Body;
@@ -58,7 +67,13 @@ namespace com.vorwardit.wordsworthcms.API
         [HttpDelete]
         public async Task<IHttpActionResult> Delete(Guid id)
         {
-            await pageLayoutService.DeleteAsync(id);
+			var allowed = await userService.HasPermission(User.Identity.GetUserId(), PermissionType.Designer, null);
+			if (!allowed)
+			{
+				return BadRequest("User is not authorised to perform this action");
+			}
+
+			await pageLayoutService.DeleteAsync(id);
             return Ok();
         }
     }
